@@ -1,6 +1,6 @@
 {
   "name": "react-admin-json-view",
-  "version": "1.2.1",
+  "version": "2.0.0",
   "description": "JSON field and input for react-admin.",
   "repository": {
     "type": "git",
@@ -27,27 +27,26 @@
     ]
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "^17.1.0",
-    "@rollup/plugin-node-resolve": "^11.1.1",
-    "@types/jest": "^26.0.20",
-    "@types/lodash": "^4.14.168",
-    "@typescript-eslint/eslint-plugin": "^4.14.2",
-    "@typescript-eslint/parser": "^4.14.2",
-    "eslint": "^7.19.0",
-    "eslint-config-prettier": "^7.2.0",
-    "eslint-plugin-prettier": "^3.3.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "jest": "^26.6.3",
-    "prettier": "^2.1.2",
-    "react": "^17.0.1",
-    "react-admin": "^3.12.0",
-    "react-dom": "^17.0.1",
-    "rollup": "^2.38.4",
+    "@rollup/plugin-commonjs": "^22.0.1",
+    "@rollup/plugin-node-resolve": "^13.3.0",
+    "@types/jest": "^28.1.6",
+    "@typescript-eslint/eslint-plugin": "^5.32.0",
+    "@typescript-eslint/parser": "^5.32.0",
+    "eslint": "^8.21.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "jest": "^28.1.3",
+    "prettier": "^2.7.1",
+    "react": "^18.2.0",
+    "react-admin": "^4.2.5",
+    "react-dom": "^18.2.0",
+    "rollup": "^2.77.2",
     "rollup-plugin-delete": "^2.0.0",
     "rollup-plugin-peer-deps-external": "^2.2.4",
-    "rollup-plugin-typescript2": "^0.29.0",
-    "ts-jest": "^26.5.0",
-    "typescript": "^4.0.5"
+    "rollup-plugin-typescript2": "^0.32.1",
+    "ts-jest": "^28.0.7",
+    "typescript": "^4.7.4"
   },
   "peerDependencies": {
     "react": "*",
