@@ -1,22 +1,21 @@
-import get from "lodash/get";
 import React from "react";
-import { FieldProps } from "react-admin";
+import { FieldProps, useRecordContext } from "react-admin";
 import ReactJson, { ReactJsonViewProps } from "react-json-view";
 
 type Props = {
   source: string;
-  record?: any;
   jsonString?: boolean;
   reactJsonOptions?: Omit<ReactJsonViewProps, "src">;
 } & FieldProps;
 
 export const JsonField: React.FC<Props> = ({
   source,
-  record = {},
   jsonString = false,
   reactJsonOptions = {},
 }) => {
-  let src = get(record, source);
+  const record = useRecordContext();
+
+  let src = record[source];
 
   if (jsonString) {
     src = JSON.parse(src);
